backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
asyncpg==0.30.0
qdrant-client==1.11.3
celery==5.4.0
redis==5.1.0
langfuse==2.42.0
httpx==0.27.2
python-multipart==0.0.9
python-magic==0.4.27
pymupdf==1.24.10
openai==1.51.0
python-docx==1.1.2

